Update readme. Add 'favorites' and 'queue' flags to Playlist. Stop returning track details (other than track id) from playlist serialization. Add handler to toggle a track as a favorite. Add polling to keep sessions alive as long as the tab is open. Create artwork directory on startup. MusicScanner will now save album art to the art folder. Last.fm is now just a fallback for album art on tracks that don't have album art. In PlaylistLogic, add a method to just add/remove a single track from a playlist. Create 'favorites' and 'queue' playlists for newly created users. Add a (barely working) search screen. Add a 'SearchTest' screen to experiment with react-virtualized. Begin adding a menu of actions to each mediaItem to do things like 'add to playlist' or 'add to favorites'. Replace bulma-extensions slider with rc-slider. Mainly to get a different color for the area to the right of the handle. Phew.

+++ src/main/java/net/ehicks/loon/handlers/PlaylistHandler.java
@@ -120,11 +120,22 @@ public class PlaylistHandler
         return playlist;
     }
 
+    @PostMapping("/toggleFavorite")
+    public Playlist add(@AuthenticationPrincipal User user, @RequestParam Long trackId)
+    {
+        Playlist playlist = playlistRepo.findByUserIdAndFavoritesTrue(user.getId());
+
+        if (playlist != null)
+            playlistLogic.addOrRemoveTrack(playlist, trackId);
+
+        return playlist;
+    }
+
     @DeleteMapping("/{playlistId}")
     public ResponseEntity delete(@AuthenticationPrincipal User user, @PathVariable long playlistId)
     {
         Playlist playlist = playlistRepo.findById(playlistId).orElse(null);
-        if (playlist != null)
+        if (playlist != null && !playlist.getFavorites() && !playlist.getQueue())
         {
             playlistLogic.setTrackIds(playlist, new ArrayList<>());
             playlistRepo.delete(playlist);
